% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EtiquetteR-package.R
\docType{package}
\name{InsectLabelR-package}
\alias{InsectLabelR}
\alias{InsectLabelR-package}
\title{InsectLabelR: Create Labels for Insect in Collection}
\description{
Streamlines the creation of high-quality labels for insect pinning. By taking a dataset as input, the package allow to generate printable labels in 'LaTeX' and PDF format, helping researchers and entomologists maintain accurate and standardized specimen records. Requires a compatible installation of 'pdflatex' (e.g. \url{https://www.tug.org/texlive/}). For enhanced accessibility, the package includes a user-friendly 'shiny' application (accessible online \url{https://nicolas-moiroux.shinyapps.io/InsectLabelR/}), which provides a graphical interface for generating labels without requiring programming expertise.
}
\author{
\strong{Maintainer}: Nicolas Moiroux \email{nicolas.moiroux@ird.fr} (\href{https://orcid.org/0000-0001-6755-6167}{ORCID})

Authors:
\itemize{
  \item Nil Rahola \email{nil.rahola@ird.fr} (\href{https://orcid.org/0000-0003-4067-6438}{ORCID})
}

}
\keyword{internal}
