% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_lab_gen.R
\name{print_bottom}
\alias{print_bottom}
\title{Append the Closing Section to a 'LaTeX' Document}
\usage{
print_bottom(file_out)
}
\arguments{
\item{file_out}{A character string specifying the path to the output file where
the 'LaTeX' closing commands will be appended.}
}
\value{
This function appends 'LaTeX' code to the file specified in \code{file_out}.
It does not return any output in R.
}
\description{
This function writes the closing commands to a 'LaTeX' document. Specifically,
it appends the end of a \verb{multicols*} environment and the \code{document} environment.
}
\details{
The function appends 'LaTeX' commands for ending a multiple-column layout
(using the \verb{multicols*} environment) and the document. It ensures the proper
closure of a 'LaTeX' document that was generated by the preceding steps.
}
