% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_lab_gen.R
\name{InsectLabelR_App}
\alias{InsectLabelR_App}
\title{Launch the InsectLabelR 'shiny' application}
\usage{
InsectLabelR_App()
}
\value{
A \strong{'shiny' application} object.
}
\description{
This function launches an interactive 'shiny' application allowing users
to use InsectLabelR with either example datasets included in the package
or user-provided data. It provides a graphical interface for generating
labels without requiring programming expertise.
}
\examples{
if(interactive()){
InsectLabelR_App()
}

}
